requirements.txt
fastapi
uvicorn
sse-starlette
python-multipart
psycopg2-binary
